refactor: StockExchangeFetcher -> StockSymbolFetcher 로 변경
