--- src/main/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/stockexchange/dto/StockSymbolDetail.java
@@ -1,6 +1,0 @@
-package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto;
-
-public record StockSymbolDetail(
-        String symbol
-) {
-}
