+++ src/main/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/stocksymbol/StockSymbolFetcher.java
@@ -1,17 +1,17 @@
-package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange;
+package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol;
 
 import com.gyeryongbrother.pickandtest.domain.core.StockExchange;
 import com.gyeryongbrother.pickandtest.infrastructure.client.FetcherSupport;
 import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.common.HeaderProvider;
 import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.common.NasdaqUrlProvider;
-import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto.StockSymbolResponse;
+import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto.StockSymbolResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class StockExchangeFetcher {
+public class StockSymbolFetcher {
 
     private final NasdaqUrlProvider nasdaqUrlProvider;
     private final HeaderProvider headerProvider;
