+++ src/main/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/stocksymbol/dto/StockSymbolData.java
@@ -1,4 +1,4 @@
-package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto;
+package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
