+++ src/main/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/stocksymbol/dto/StockSymbolDetail.java
@@ -1,0 +1,6 @@
+package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto;
+
+public record StockSymbolDetail(
+        String symbol
+) {
+}
