+++ src/test/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/NasdaqClientTest.java
@@ -1,13 +1,13 @@
 package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq;
 
 import static com.gyeryongbrother.pickandtest.domain.core.StockExchange.NASDAQ;
-import static com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto.StockSymbolResponseFixture.stockSymbolResponse;
+import static com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto.StockSymbolResponseFixture.stockSymbolResponse;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.any;
 import static org.mockito.BDDMockito.given;
 
-import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.StockExchangeFetcher;
-import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto.StockSymbolResponse;
+import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.StockSymbolFetcher;
+import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto.StockSymbolResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -20,13 +20,13 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class NasdaqClientTest {
 
     @Mock
-    private StockExchangeFetcher stockExchangeFetcher;
+    private StockSymbolFetcher stockSymbolFetcher;
 
     private NasdaqClient nasdaqClient;
 
     @BeforeEach
     void setUp() {
-        nasdaqClient = new NasdaqClient(stockExchangeFetcher);
+        nasdaqClient = new NasdaqClient(stockSymbolFetcher);
     }
 
     @Test
@@ -34,7 +34,7 @@ class NasdaqClientTest {
     void fetchStockExchange() {
         // given
         StockSymbolResponse stockSymbolResponse = stockSymbolResponse();
-        given(stockExchangeFetcher.fetchStockSymbol(any()))
+        given(stockSymbolFetcher.fetchStockSymbol(any()))
                 .willReturn(stockSymbolResponse);
 
         // when
