+++ src/test/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/stocksymbol/StockSymbolFetcherTest.java
@@ -1,7 +1,7 @@
-package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange;
+package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol;
 
 import static com.gyeryongbrother.pickandtest.domain.core.StockExchange.NASDAQ;
-import static com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto.StockSymbolResponseFixture.stockSymbolResponse;
+import static com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto.StockSymbolResponseFixture.stockSymbolResponse;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.anyString;
@@ -10,7 +10,7 @@ import static org.mockito.BDDMockito.given;
 import com.gyeryongbrother.pickandtest.infrastructure.client.FetcherSupport;
 import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.common.HeaderProvider;
 import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.common.NasdaqUrlProvider;
-import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto.StockSymbolResponse;
+import com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto.StockSymbolResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,28 +21,28 @@ import org.springframework.http.ResponseEntity;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayName("주식 티커들을 가져온다")
-class StockExchangeFetcherTest {
+class StockSymbolFetcherTest {
 
     @Mock
     private FetcherSupport fetcherSupport;
 
-    private StockExchangeFetcher stockExchangeFetcher;
+    private StockSymbolFetcher stockSymbolFetcher;
 
     @BeforeEach
     void setUp() {
-        stockExchangeFetcher = new StockExchangeFetcher(new NasdaqUrlProvider(), new HeaderProvider(), fetcherSupport);
+        stockSymbolFetcher = new StockSymbolFetcher(new NasdaqUrlProvider(), new HeaderProvider(), fetcherSupport);
     }
 
     @Test
     @DisplayName("주식 티커들을 가져온다")
-    void fetchStockExchange() {
+    void fetchStockSymbol() {
         // given
         StockSymbolResponse stockSymbolResponse = stockSymbolResponse();
         given(fetcherSupport.get(anyString(), any(), any()))
                 .willReturn(ResponseEntity.ok(stockSymbolResponse));
 
         // when
-        StockSymbolResponse result = stockExchangeFetcher.fetchStockSymbol(NASDAQ);
+        StockSymbolResponse result = stockSymbolFetcher.fetchStockSymbol(NASDAQ);
 
         // then
         assertThat(result).isEqualTo(stockSymbolResponse);
