+++ src/test/java/com/gyeryongbrother/pickandtest/infrastructure/client/nasdaq/stocksymbol/dto/StockSymbolResponseFixture.java
@@ -1,4 +1,4 @@
-package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stockexchange.dto;
+package com.gyeryongbrother.pickandtest.infrastructure.client.nasdaq.stocksymbol.dto;
 
 import java.util.List;
